#337: Introduce a byte array abstract stack item to consolidate code for Buffer and ByteString item types.
